bch_bindgen: fix clippy errors and warnings
Mark printbuf methods that take raw pointers as unsafe (sb_to_text,
sb_to_text_with_names, prt_bitflags), remove unnecessary u64 casts in
bitmask.rs, use div_ceil in journal sector calculation, and suppress
should_implement_trait on BtreeNodeIter::next which intentionally
returns Result.

+++ bch_bindgen/src/data/extents.rs
@@ -10,7 +10,7 @@ include!(concat!(env!("OUT_DIR"), "/extent_entry_types_gen.rs"));
 ///
 /// Returns `u32::MAX` if the type field is zero (invalid).
 pub fn extent_entry_type(entry: &c::bch_extent_entry) -> u32 {
-    let t = unsafe { entry.type_ } as u64;
+    let t = unsafe { entry.type_ };
     if t != 0 { t.trailing_zeros() } else { u32::MAX }
 }
